Cap2 Terrell Driving ty Rich

+++ java/tenmo-server/src/main/java/com/techelevator/tenmo/dao/UserDAO.java
@@ -2,6 +2,7 @@ package com.techelevator.tenmo.dao;
 
 import com.techelevator.tenmo.model.User;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface UserDAO {
@@ -14,5 +15,5 @@ public interface UserDAO {
 
     boolean create(String username, String password);
     
-    public double getUserBalance(String username);
+    public BigDecimal getUserBalance(String username);
 }
